refactor: rename the task 2 file
