+++ game/tasks/task01/srp_introduction.py
@@ -20,7 +20,7 @@ import arcade
 
 from ...core.registration import Register, SpriteRegister
 from ...core.views.rpg_sprites import CharacterSprite
-from ..task02.task_o_inventory import PlayerSprite
+from ..task02.task_o_equipped_item import PlayerSprite
 from ..task03.task_l_crafting import PlayerInventory
 from .task_s_select_character import SOURCE_NAME
